refactor : CarValidator 구현 로직을 Car 로 통합
