+++ src/main/java/racingcar/domain/Car.java
@@ -1,11 +1,13 @@
 package racingcar.domain;
 
-import racingcar.util.validator.CarValidator;
-
 public class Car implements Comparable<Car> {
 
 	private static final int INITIAL_POSITION = 0;
 	private static final int STANDARD_OF_MOVING = 4;
+	private static final int NAME_LENGTH_LIMIT = 5;
+	private static final int MINIMUM_POSITION = 0;
+	private static final String INVALID_CAR_NAME_ERROR_MESSAGE = "자동차의 이름은 1~5글자여야 합니다.";
+	private static final String INVALID_POSITION_ERROR_MESSAGE = "자동차 위치는 0 이상이여야 합니다.";
 
 	private final String name;
 	private int position;
@@ -16,14 +18,14 @@ public class Car implements Comparable<Car> {
 	}
 
 	public static Car of(String name) {
-		CarValidator.checkNameLength(name);
+		checkNameLength(name);
 
 		return new Car(name, INITIAL_POSITION);
 	}
 
 	public static Car of(String name, int position) {
-		CarValidator.checkNameLength(name);
-		CarValidator.checkPositionRange(position);
+		checkNameLength(name);
+		checkPositionRange(position);
 
 		return new Car(name, position);
 	}
@@ -38,6 +40,21 @@ public class Car implements Comparable<Car> {
 
 	public Car copy() {
 		return Car.of(this.name, this.position);
+	}
+	public static void checkNameLength(String name) {
+		if (name.isEmpty() || isOverLength(name)) {
+			throw new IllegalArgumentException(INVALID_CAR_NAME_ERROR_MESSAGE);
+		}
+	}
+
+	private static boolean isOverLength(String name) {
+		return name.length() > NAME_LENGTH_LIMIT;
+	}
+
+	public static void checkPositionRange(int position) {
+		if(position < MINIMUM_POSITION) {
+			throw new IllegalArgumentException(INVALID_POSITION_ERROR_MESSAGE);
+		}
 	}
 
 	public void move(int movingValue) {
